Move getFlagType to utils
